requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
albumentations==1.3.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
dominate==2.7.0
efficientnet-pytorch==0.7.1
einops==0.6.0
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
huggingface-hub==0.11.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.1.0
importlib-resources==5.12.0
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.4
kornia==0.6.8
lion-pytorch==0.0.7
lpips==0.1.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.0
multidict==6.0.2
munch==2.5.0
networkx==2.8.8
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.3
pathtools==0.1.2
Pillow==9.3.0
pretrainedmodels==0.7.4
promise==2.3
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pyDeprecate==0.3.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.2
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.12
torch==1.11.0+cu113
torchmetrics==0.6.0
torchvision==0.12.0+cu113
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
ultralytics==8.0.58
urllib3==1.26.13
visdom==0.2.4
vit-pytorch==1.2.0
wandb==0.14.0
websocket-client==1.5.1
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.10.0

anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2022.9.24
cffi==1.15.1
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
executing==1.2.0
fastjsonschema==2.16.3
fqdn==1.5.1
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.2.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyrsistent==0.19.3
python-json-logger==2.0.7
pytz==2022.7.1
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.0
setuptools==59.5.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
traitlets==5.9.0
uri-template==1.2.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
wheel==0.37.1
widgetsnbextension==4.0.7
pip==22.2.2